Moving content auditor to its own file
